refactor(file structure): extract folder feature-reading-list
