+++ feat-popup/index.css
@@ -1,4 +1,4 @@
-@import url(../modules/reading-list.css);
+@import url(../feature-reading-list/reading-list.css);
 @import "../modules/status-bar.css";
 
 * {
