+++ feat-popup/index.js
@@ -1,5 +1,5 @@
 import '../modules/prototype.mjs'
-import * as readingList from '../modules/readingList.mjs'
+import * as readingList from '../feature-reading-list/readingList.mjs'
 import * as statusBar from '../modules/statusBar.js'
 
 $(async () => {
